move app database to core instead of app
