+++ splash/src/main/res/layout/splash_fragment.xml
@@ -21,6 +21,7 @@
         android:id="@+id/progressSplash"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:visibility="gone"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
